+++ test_roadmap.py
@@ -15,14 +15,14 @@ class TestRoadmapFunctions(unittest.TestCase):
     def setUp(self):
         # this folder is used for storing data during test
         self.test_folder = "tests/test_folder"
-        # this file did not excist
+        # this file did not exist
         self.test_file = "test_file.yml"
-        # this is a excisting roadmap
-        self.test_excisting_file = "tests/roadmap.yml"
-        # this is the version of the excisting roadmap
-        # if you modifiy this file, make shure to modifiy his version
-        # version is calculated using md5 and take the first and last 4 characters as version
-        self.version_excisting_roadmap = "880a29cf"
+        # this is an existing roadmap
+        self.test_existing_file = "tests/roadmap.yml"
+        # this is the version of the existing roadmap
+        # if you modify this file, make shure to modify his version.
+        # version is calculated using md5 and take the first and last 4 characters
+        self.version_existing_roadmap = "880a29cf"
 
     def test_create_output_folder(self):
         # Test when the folder does not exist
@@ -58,15 +58,15 @@ class TestRoadmapFunctions(unittest.TestCase):
     def test_roadmap_yml_version_id(self):
         # test if we get none for non-existing file
         self.assertIsNone(calculate_roadmap_version(self.test_file))
-        # test if the version id of roadmap fullfills our expectations
-        self.assertEqual(calculate_roadmap_version(self.test_excisting_file), self.version_excisting_roadmap)
+        # test if the version id of roadmap fulfills our expectations
+        self.assertEqual(calculate_roadmap_version(self.test_existing_file), self.version_existing_roadmap)
 
-    def test_preconditions_in_test_excisting_file(self):
+    def test_preconditions_in_test_exciting_file(self):
         # this test is to check if your test.yml has some predefined conditions for testing
         #
         #   the conditions are checked twice: once for version id, which is md5sum of file
-        #   and second for the condtions here
-        project = dict(read_roadmap_definition(self.test_excisting_file))
+        #   and second for the conditions here
+        project = dict(read_roadmap_definition(self.test_existing_file))
         # did we have milestones
         self.assertIn("milestones", project)
         # check if objectives are in project
@@ -76,12 +76,12 @@ class TestRoadmapFunctions(unittest.TestCase):
 
         # check if first milestone contains deliverables
         self.assertIn("deliverables", project["milestones"][0])
-        # check if first objecttive contains keyresults
+        # check if first objective contains keyresults
         self.assertIn("keyresults", project["objectives"][0])
 
         # check if first milestone first deliverables contains well known quantifiers
         self.assertIn("quantifiers", project["milestones"][0]['deliverables'][0])
-        # check if first objecttive first keyresult contains well known quantifiers
+        # check if first objective first keyresult contains well known quantifiers
         self.assertIn("quantifiers", project["objectives"][0]['keyresults'][0])
 
         # check conditions for cost of delay and wsjf calculation
@@ -101,16 +101,16 @@ class TestRoadmapFunctions(unittest.TestCase):
 
     def test_is_level0_element_removed_from_project(self):
         remove_pattern = "milestones"
-        # test if a element from project is scipped during operation
-        project = read_roadmap_definition(self.test_excisting_file)
+        # test if an element from project is skipped during operation
+        project = read_roadmap_definition(self.test_existing_file)
         remove_element(remove_pattern, project=project)
         self.assertNotIn("milestones", project)
         self.assertIn("objectives", project)
 
     def test_is_level1_element_removed_from_project(self):
         remove_pattern = "milestones.deliverables"
-        # test if a element from project is scipped during operation
-        project = dict(read_roadmap_definition(self.test_excisting_file))
+        # test if an element from project is skipped during operation
+        project = dict(read_roadmap_definition(self.test_existing_file))
         # remove milestones deliverables
         remove_element(remove_pattern, project=project)
         # check if deliverables from first milestone is removed
@@ -122,8 +122,8 @@ class TestRoadmapFunctions(unittest.TestCase):
 
     def test_is_level2_element_removed_from_project(self):
         remove_pattern = "milestones.deliverables.todos"
-        # test if a element from project is scipped during operation
-        project = dict(read_roadmap_definition(self.test_excisting_file))
+        # test if an element from project is skipped during operation
+        project = dict(read_roadmap_definition(self.test_existing_file))
         # remove milestones.deliverables.todos
         remove_element(remove_pattern, project=project)
         # check if objectives are still present
@@ -137,8 +137,8 @@ class TestRoadmapFunctions(unittest.TestCase):
 
     def test_is_level1_description_removed_from_project(self):
         remove_pattern = "milestones.description"
-        # test if a element from project is scipped during operation
-        project = dict(read_roadmap_definition(self.test_excisting_file))
+        # test if an element from project is skipped during operation
+        project = dict(read_roadmap_definition(self.test_existing_file))
         # check if milestones are in project
         self.assertIn("milestones", project)
         # check if objectives are in project
@@ -156,8 +156,8 @@ class TestRoadmapFunctions(unittest.TestCase):
 
     def test_is_level3_removed_from_project(self):
         remove_pattern = "milestones.deliverables.todos.description"
-        # test if a element from project is scipped during operation
-        project = dict(read_roadmap_definition(self.test_excisting_file))
+        # test if an element from project is skipped during operation
+        project = dict(read_roadmap_definition(self.test_existing_file))
         # check if milestones are in project
         self.assertIn("milestones", project)
         # check if objectives are in project
@@ -174,13 +174,13 @@ class TestRoadmapFunctions(unittest.TestCase):
         self.assertIn("milestones", project)
         # check if milestones first deliverable is removed
         self.assertNotIn("description", project["milestones"][0]['deliverables'][0]['todos'][0])
-        # check if objectives first keyresult todo element contains still description
+        # check if objectives first keyresult todos-element contains still description
         self.assertIn("description", project["objectives"][0]['keyresults'][0]['todos'][0])
 
     def test_is_level3_string_element_removal_is_set_to_none(self):
         remove_pattern = "milestones.deliverables.quantifiers.jobsize"
-        # test if a element from project is scipped during operation
-        project = dict(read_roadmap_definition(self.test_excisting_file))
+        # test if an element from project is skipped during operation
+        project = dict(read_roadmap_definition(self.test_existing_file))
         # check if milestones first deliverables quantifiers element contains jobsize
         self.assertIn("jobsize", project["milestones"][0]['deliverables'][0]['quantifiers'])
         # remove quantifiers.jobsize
@@ -188,10 +188,10 @@ class TestRoadmapFunctions(unittest.TestCase):
         # check if milestones first deliverable is None
         self.assertIsNone(project["milestones"][0]['deliverables'][0]['quantifiers']['jobsize'])
 
-    def test_correct_errorhandling_for_unsupported_remove_string(self):
+    def test_correct_error_handling_for_unsupported_remove_string(self):
         # test if we handle all errors correctly
         remove_pattern = "."
-        project = dict(read_roadmap_definition(self.test_excisting_file))
+        project = dict(read_roadmap_definition(self.test_existing_file))
         with self.assertRaises(ValueError):
             remove_element(remove_pattern, project=project)
         # check if objectives are still present
@@ -200,8 +200,10 @@ class TestRoadmapFunctions(unittest.TestCase):
         self.assertIn("milestones", project)
 
     def test_remove_handles_multiple_levels_to_multiple_times(self):
-        skip_items = "milestones.deliverables.quantifiers.weighted_shortest_job_first,objectives.keyresults.todos,objectives.keyresults,milestones"
-        project = dict(read_roadmap_definition(self.test_excisting_file))
+        skip_items = ("milestones.deliverables.quantifiers.weighted_shortest_job_first,"
+                      "objectives.keyresults.todos,"
+                      "objectives.keyresults,milestones")
+        project = dict(read_roadmap_definition(self.test_existing_file))
         # check if milestones are present
         self.assertIn("milestones", project)
         # first round
@@ -228,10 +230,10 @@ class TestRoadmapFunctions(unittest.TestCase):
         # check for valid input with 10 for each value should give 30
         self.assertEqual(calculate_cost_of_delay(user_business_value=10, time_criticality=10,
                                                  opportunity_enablement_or_risk_reduction=10), 30)
-        # check for handling non valid input
+        # check for handling non-valid input
 
     def test_calculate_cost_of_delay_non_valid_inputs(self):
-        # test for correct handling non valid inputs during calculation of cost_of_delay
+        # test for correct handling non-valid inputs during calculation of cost_of_delay
         with self.assertRaises(ValueError):
             calculate_cost_of_delay(user_business_value=0, time_criticality=0,
                                     opportunity_enablement_or_risk_reduction=11)
@@ -258,10 +260,10 @@ class TestRoadmapFunctions(unittest.TestCase):
         self.assertEqual(calculate_weighted_shortest_job_first(cost_of_delay=30, jobsize=10), 3.00)
         # check for valid input with floating result 
         self.assertEqual(calculate_weighted_shortest_job_first(cost_of_delay=16, jobsize=3), 5.33)
-        # check for handling non valid input
+        # check for handling non-valid input
 
     def test_calculate_weighted_shortest_job_first_non_valid_inputs(self):
-        # test for correct handling non valid inputs during calculation of wsjf
+        # test for correct handling non-valid inputs during calculation of wsjf
         with self.assertRaises(ValueError):
             calculate_weighted_shortest_job_first(cost_of_delay="16", jobsize="3")
             calculate_weighted_shortest_job_first(cost_of_delay=19)
@@ -272,7 +274,7 @@ class TestRoadmapFunctions(unittest.TestCase):
 
     def test_calculate_wsjf_quantifiers_if_weighted_shortest_job_is_set(self):
         # test if we calculate correctly and quantifier will be added to project
-        project = dict(read_roadmap_definition(self.test_excisting_file))
+        project = dict(read_roadmap_definition(self.test_existing_file))
         # set some weird WSJF
         weird_wsjf = 100.01
         # store in project
@@ -288,7 +290,7 @@ class TestRoadmapFunctions(unittest.TestCase):
 
     def test_calculate_wsjf_quantifiers_if_weighted_shortest_job_is_not_set(self):
         # test if we calculate correctly and quantifier will be added to project
-        project = dict(read_roadmap_definition(self.test_excisting_file))
+        project = dict(read_roadmap_definition(self.test_existing_file))
         # set user_business_value to None
         project["objectives"][0]['keyresults'][0]['quantifiers']['user_business_value'] = None
         # calculate
@@ -308,7 +310,7 @@ class TestRoadmapFunctions(unittest.TestCase):
 
     def test_calculate_wsjf_quantifiers_if_cost_of_delay_is_set(self):
         # test if we calculate correctly and quantifier will be added to project
-        project = dict(read_roadmap_definition(self.test_excisting_file))
+        project = dict(read_roadmap_definition(self.test_existing_file))
         # set some weird cost_of_delay
         weird_cost_of_delay = 100
         project["objectives"][0]['keyresults'][0]['quantifiers']['cost_of_delay'] = weird_cost_of_delay
@@ -323,7 +325,7 @@ class TestRoadmapFunctions(unittest.TestCase):
 
     def test_calculate_wsjf_quantifiers_for_milestones_deliverables(self):
         # test if we calculate correctly and quantifier will be added to project
-        project = dict(read_roadmap_definition(self.test_excisting_file))
+        project = dict(read_roadmap_definition(self.test_existing_file))
         project["milestones"][0]['deliverables'] = calculate_wsjf_quantifiers_for_element_items(
             project["milestones"][0]['deliverables'])
         # did we still have wsjf?
@@ -345,7 +347,7 @@ class TestRoadmapFunctions(unittest.TestCase):
 
     def test_calculate_wsjf_quantifiers_for_objectives_keyresult(self):
         # test if we calculate correctly and quantifier will be added to project
-        project = dict(read_roadmap_definition(self.test_excisting_file))
+        project = dict(read_roadmap_definition(self.test_existing_file))
         project["milestones"][0]['deliverables'] = calculate_wsjf_quantifiers_for_element_items(
             project["objectives"][0]['keyresults'])
         # did we still have wsjf?
@@ -364,7 +366,7 @@ class TestRoadmapFunctions(unittest.TestCase):
 
     def test_get_key_value_list(self):
         # test if we build a key-value list correctly
-        project = dict(read_roadmap_definition(self.test_excisting_file))
+        project = dict(read_roadmap_definition(self.test_existing_file))
         project_as_list = get_key_value_list(element=project)
         # project_as_list is list?
         self.assertIsInstance(project_as_list, list)
@@ -388,13 +390,13 @@ class TestRoadmapFunctions(unittest.TestCase):
         self.assertEqual(project_as_list[0]['value'], "M1")
         # check for error handling
         project_as_list = get_key_value_list()
-        # we always get a empty key, value list
+        # we always get an empty key, value list
         self.assertIsNone(project_as_list[0]['key'])
         self.assertIsNone(project_as_list[0]['value'])
 
     def test_get_filtered_key_value_list(self):
         # test if we filter a key-value list correctly
-        project = dict(read_roadmap_definition(self.test_excisting_file))
+        project = dict(read_roadmap_definition(self.test_existing_file))
         # first, test with a known list
         project_as_list = get_key_value_list(element=project)
         filtered_list = get_filtered_key_value_list(key_value_list=project_as_list, filter_for_keys="milestones.title",